+++ src/GamePlanetarium.Domain/Statistics/GameStatisticsDataCollector.cs
@@ -6,8 +6,8 @@ namespace GamePlanetarium.Domain.Statistics;
 public class GameStatisticsDataCollector
 {
     public bool IsGameEnded { get; private set; }
-    public Dictionary<int, QuestionStatisticsData> QuestionsStatistics { get; } = new();
-    public int QuestionsAnsweredCount => QuestionsStatistics.Count;
+    public Dictionary<byte, QuestionStatisticsData> QuestionsStatistics { get; } = new();
+    public byte QuestionsAnsweredCount => (byte)QuestionsStatistics.Count;
     public DateOnly DateStamp { get; }
 
     public GameStatisticsDataCollector(IGameObservable game, DateOnly? dateStamp = null)
@@ -21,13 +21,13 @@ public class GameStatisticsDataCollector
 
     private void OnQuestionAnswered(object? sender, AnsweredQuestionInfo info)
     {
-        int questionNumber = info.QuestionNumber;
+        byte questionNumber = info.QuestionNumber;
         if (!QuestionsStatistics.ContainsKey(questionNumber))
         {
             QuestionsStatistics.Add(questionNumber, new QuestionStatisticsData
             {
                 FirstAnswerText = info.FirstAnswer.Text,
-                QuestionOrder = QuestionsAnsweredCount + 1,
+                QuestionOrder = (byte)(QuestionsAnsweredCount + 1)
             });
         }
         QuestionsStatistics[questionNumber].UpdateIncorrectAnswersCount(info.IsAnsweredCorrectly);
